requirements.txt
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
